+++ pkg/deployments/README.md
@@ -91,8 +91,8 @@ Returns an object with all contracts from a deployment and their addresses.
 | Batch Relayer V3                                        | [`20220720-batch-relayer-v3`](./tasks/20220720-batch-relayer-v3)                                     |
 | Swap, join and exit simulations (queries)               | [`20220721-balancer-queries`](./tasks/20220721-balancer-queries)                                     |
 | Protocol fee percentages provider                       | [`20220725-protocol-fee-percentages-provider`](./tasks/20220725-protocol-fee-percentages-provider)   |
-| Linear Pools for Aave aTokens with built-in rebalancing | [`20220729-aave-rebalanced-linear-pool`](./tasks/20220729-aave-rebalanced-linear-pool)               |
 | Child Chain Gauge Reward Helper                         | [`20220812-child-chain-reward-helper`](./tasks/20220812-child-chain-reward-helper)                   |
+| Linear Pools for Aave aTokens with built-in rebalancing | [`20220817-aave-rebalanced-linear-pool`](./tasks/20220817-aave-rebalanced-linear-pool)               |
 
 ## Scripts
