+++ pkg/deployments/deployment-txs/arbitrum.json
@@ -31,5 +31,6 @@
   "0x4f4269F7008430860693aeDB86b0C55962D6E9b7": "0x19b1b189804af1d2d7b516846ed1bed81936760bdde4a36b74811501c697bbd6",
   "0xE39B5e3B6D74016b2F6A9673D7d7493B6DF549d5": "0x710d93aab52b6c10197eab20f9d6db1af3931f9890233d8832268291ef2f54b3",
   "0x5ef4c5352882b10893b70DbcaA0C000965bd23c5": "0xbc9d8127b5c39816b24486a35b8ef9eb3687d8e343f963af5a4feaed34fdc5fd",
-  "0xA0DAbEBAAd1b243BBb243f933013d560819eB66f": "0xefb8d9da6daf08dd54a337d9bbf053217bc06077eecfca03e8571abb0d2c20df"
+  "0xA0DAbEBAAd1b243BBb243f933013d560819eB66f": "0xefb8d9da6daf08dd54a337d9bbf053217bc06077eecfca03e8571abb0d2c20df",
+  "0xe2E901AB09f37884BA31622dF3Ca7FC19AA443Be": "0xec419da1095c1b65955b8ac18a1ebd2152d4cf354cd3dd54a7f3d6f7785bcc39"
 }
