+++ pkg/deployments/deployment-txs/goerli.json
@@ -53,5 +53,6 @@
   "0x8F42aDBbA1B16EaAE3BB5754915E0D06059aDd75": "0x3254ef870ff30a8affc11de1408ef686798c2dc7796ad95803bd30ccdba52e0d",
   "0x0F3e0c4218b7b0108a3643cFe9D3ec0d4F57c54e": "0xc2a0b14c5dac2b550e98f6c07f7fabed073fd0cacd7536910191152189ffd07f",
   "0xD7FAD3bd59D6477cbe1BE7f646F7f1BA25b230f8": "0x50dc987d79951ea16cb07ce44189c1a51253454fb7a6f50290f84709b880ac17",
-  "0xC128a9954e6c874eA3d62ce62B468bA073093F25": "0x75d95195e938dd905b7be5f26447dafcadf9442fa3a54e03937595265f506320"
+  "0xC128a9954e6c874eA3d62ce62B468bA073093F25": "0x75d95195e938dd905b7be5f26447dafcadf9442fa3a54e03937595265f506320",
+  "0xC128468b7Ce63eA702C1f104D55A2566b13D3ABD": "0xecf621288b707681d22c618755d93c8abfab7fa75aef9921785591df86e9b04f"
 }
