+++ pkg/deployments/deployment-txs/mainnet.json
@@ -57,5 +57,6 @@
   "0xD966d712F470067B60D37246404D6DFe5Bf0B419": "0x5bafe8ecf91630d92c1066c87ef30130b455535e00d74282a18328ddb30643bc",
   "0xE39B5e3B6D74016b2F6A9673D7d7493B6DF549d5": "0x30799534f3a0ab8c7fa492b88b56e9354152ffaddad15415184a3926c0dd9b09",
   "0x97207B095e4D5C9a6e4cfbfcd2C3358E03B90c4A": "0xb903f8ba66f446c11d9a44c3bfd27a39dfe4e27e3e2ba8a88cf3e73625114862",
-  "0x657dEe983987E1A79A996D972fE68411AeF05e8D": "0xaf83b4bebc6a276fd0579ce281ac7ad86459a35fefa2627c9a56a98ba0e4b997"
+  "0x657dEe983987E1A79A996D972fE68411AeF05e8D": "0xaf83b4bebc6a276fd0579ce281ac7ad86459a35fefa2627c9a56a98ba0e4b997",
+  "0x6A0AC04f5C2A10297D5FA79FA6358837a8770041": "0xa20d7dc195ecdce2cf1870351872d4322a59d5e8467134d41c9815eb41e81761"
 }
