+++ pkg/deployments/deployment-txs/optimism.json
@@ -22,5 +22,6 @@
   "0x5c89d9E1A668d0B39F4EC242F280bdDd445C633c": "0x79c7d4e9bb3cdf24d9cd2e32813255a978e05c68e25615615f8756d1a652e0a7",
   "0xE39B5e3B6D74016b2F6A9673D7d7493B6DF549d5": "0xf3b2aaf3e12c7de0987dc99a26242b227b9bc055342dda2e013dab0657d6f9f1",
   "0xacAaC3e6D6Df918Bf3c809DFC7d42de0e4a72d4C": "0x92cb8b071f59c0f292aa878b2d236d0c602c07bbc72cd0511b0b7c7bf3ee6142",
-  "0x8aB784368A1883DA90D8513b48801e2Db1cb2D5D": "0xc1259148af7bf9e61ca86191a8f5b064f19c070799758f8b9b450643ba11de65"
+  "0x8aB784368A1883DA90D8513b48801e2Db1cb2D5D": "0xc1259148af7bf9e61ca86191a8f5b064f19c070799758f8b9b450643ba11de65",
+  "0x994086630773dC6cB54D3A5E0Ef0963532789E75": "0x4f161e34f6147e58817bc80cd72e2d2d5441de5a0e5cf6b2ccbc522299d0f37c"
 }
