+++ pkg/deployments/deployment-txs/polygon.json
@@ -31,5 +31,6 @@
   "0x5ef4c5352882b10893b70DbcaA0C000965bd23c5": "0x4b4ddc4ceafa144df2f7eb88290aaa24db8c7922d68bb63439895fc52eaf83bf",
   "0xE39B5e3B6D74016b2F6A9673D7d7493B6DF549d5": "0x0b74f5c230f9b7df8c7a7f0d1ebd5e6c3fab51a67a9bcc8f05c350180041682e",
   "0x42AC0e6FA47385D55Aff070d79eF0079868C48a6": "0x1a6f3cc3e15a628eaf323b5d7a3578d497ce8fcfb8952273d20d8084621fb31f",
-  "0xaEb406b0E430BF5Ea2Dc0B9Fe62E4E53f74B3a33": "0x13e145eb41511f3a9b0f4777fa60af0e17b2ab2122b68b3abcf57263e0a5c551"
+  "0xaEb406b0E430BF5Ea2Dc0B9Fe62E4E53f74B3a33": "0x13e145eb41511f3a9b0f4777fa60af0e17b2ab2122b68b3abcf57263e0a5c551",
+  "0x8df6EfEc5547e31B0eb7d1291B511FF8a2bf987c": "0x92bed1d00f3d5b9af1ff01c5a75b766da20edc3a081507b20d20371d39100d53"
 }
