--- pkg/deployments/tasks/20220729-aave-rebalanced-linear-pool/input.ts
@@ -1,14 +1,0 @@
-import Task, { TaskMode } from '../../src/task';
-
-export type AaveLinearPoolDeployment = {
-  Vault: string;
-  BalancerQueries: string;
-};
-
-const Vault = new Task('20210418-vault', TaskMode.READ_ONLY);
-const BalancerQueries = new Task('20220721-balancer-queries', TaskMode.READ_ONLY);
-
-export default {
-  Vault,
-  BalancerQueries,
-};
