--- pkg/deployments/tasks/20220729-aave-rebalanced-linear-pool/output/goerli.json
@@ -1,4 +1,0 @@
-{
-  "AaveLinearPoolFactory": "0x8F42aDBbA1B16EaAE3BB5754915E0D06059aDd75",
-  "timestamp": 1660178675131
-}
