--- pkg/deployments/tasks/20220729-aave-rebalanced-linear-pool/readme.md
@@ -1,13 +1,0 @@
-# 2022-07-27 - Aave Rebalanced Linear Pool
-
-Deployment of the `AaveLinearPoolFactory`, for Linear Pools with a wrapped aToken. This new deployment includes a permissionless rebalancing contract that has Asset Manager privileges.
-
-## Useful Files
-
-- [Ethereum mainnet addresses](./output/mainnet.json)
-- [Polygon mainnet addresses](./output/polygon.json)
-- [Arbitrum mainnet address](./output/arbitrum.json)
-- [Optimism mainnet address](./output/optimism.json)
-- [`AaveLinearPool` ABI](./abi/AaveLinearPool.json)
-- [`AaveLinearPoolFactory` ABI](./abi/AaveLinearPoolFactory.json)
-- [`AaveLinearPoolRebalancer` ABI](./abi/AaveLinearPoolRebalancer.json)
