+++ pkg/deployments/tasks/20220817-aave-rebalanced-linear-pool/output/arbitrum.json
@@ -1,0 +1,4 @@
+{
+  "AaveLinearPoolFactory": "0xe2E901AB09f37884BA31622dF3Ca7FC19AA443Be",
+  "timestamp": 1660744054634
+}
