+++ pkg/deployments/tasks/20220817-aave-rebalanced-linear-pool/output/goerli.json
@@ -1,0 +1,4 @@
+{
+  "AaveLinearPoolFactory": "0xC128468b7Ce63eA702C1f104D55A2566b13D3ABD",
+  "timestamp": 1660743844863
+}
