+++ pkg/deployments/tasks/20220817-aave-rebalanced-linear-pool/output/mainnet.json
@@ -1,0 +1,4 @@
+{
+  "AaveLinearPoolFactory": "0x6A0AC04f5C2A10297D5FA79FA6358837a8770041",
+  "timestamp": 1660744487230
+}
