+++ pkg/deployments/tasks/20220817-aave-rebalanced-linear-pool/output/optimism.json
@@ -1,0 +1,4 @@
+{
+  "AaveLinearPoolFactory": "0x994086630773dC6cB54D3A5E0Ef0963532789E75",
+  "timestamp": 1660744343779
+}
