+++ pkg/deployments/tasks/deprecated/20211208-aave-linear-pool/readme.md
@@ -2,7 +2,7 @@
 
 > ⚠️ **DEPRECATED: do not use** ⚠️
 >
-> This factory and associated Pools have been deprecated in favor of this [updated version](../../20220729-aave-rebalanced-linear-pool), which features a permissionless Rebalancer with Asset Manager privileges.
+> This factory and associated Pools have been deprecated in favor of this [updated version](../../20220817-aave-rebalanced-linear-pool), which features a permissionless Rebalancer with Asset Manager privileges.
 
 Deployment of the `AaveLinearPoolFactory`, for Linear Pools with a wrapped aToken.
